requirements.txt
Django==4.2.7
paho-mqtt==1.6.1
python-decouple==3.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
whitenoise==6.6.0
psycopg2-binary==2.9.9